photo_server/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.0.6
djangorestframework==3.13.0
idna==3.10
pillow==11.3.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
gunicorn==23.0.0
celery
redis
django-celery-results
drf-yasg==1.21.10
